remove StaleEntityCheckpointStateBase
